requirements.txt
Flask==2.0.1
boto3==1.17.49
Pillow==8.2.0
